compiles

+++ stack/stack_rev_rotate.c
@@ -6,11 +6,11 @@
 /*   By: tunsal <tunsal@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 15:12:53 by tunsal            #+#    #+#             */
-/*   Updated: 2023/11/15 15:13:05 by tunsal           ###   ########.fr       */
+/*   Updated: 2023/11/18 03:49:55 by tunsal           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "push_swap.h"
+#include "../push_swap.h"
 
 /*
 	Shift all elements of the stack down by 1.
@@ -18,8 +18,8 @@
 */
 void	stack_rev_rotate(t_stack *s)
 {
-	size_t	i;
-	int		old_bottom;
+	int	i;
+	int	old_bottom;
 
 	old_bottom = s->data[0];
 	i = 0;
